# agent-team-workbench/contracts/web/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Team Workbench Web API
  version: 0.1.0
  description: |
    浏览器 ↔ Go Control Plane 协议（协议文档 §5）。
    约定：/api/v1 前缀；资源 ID 带类型前缀（ws_/agent_/wi_/run_/approval_/artifact_）；
    JSON 字段 snake_case；写命令要求 Idempotency-Key；可变资源带 version/ETag；
    错误一律 application/problem+json。
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /workspaces/{workspace_id}:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    patch:
      tags: [workspace]
      operationId: updateWorkspace
      description: 更新名称、时区和策略；Owner/Admin。
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkspacePatch" } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Workspace" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /workspaces/{workspace_id}/bootstrap:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [workspace]
      operationId: bootstrapWorkspace
      summary: 首屏 dashboard、agents、board、health 与 event_cursor 一次返回
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Bootstrap" } } } }
  /workspaces/{workspace_id}/dashboard:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [workspace]
      operationId: getDashboard
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Dashboard" } } } }
  /workspaces/{workspace_id}/events:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [events]
      operationId: subscribeEvents
      summary: SSE 事件流；Last-Event-ID 补发；cursor 过期返回 410
      responses:
        "200":
          description: text/event-stream
          content: { text/event-stream: { schema: { type: string } } }
        "410": { $ref: "#/components/responses/CursorExpired" }
  /workspaces/{workspace_id}/agent-profiles:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [agent-profiles]
      operationId: listAgentProfiles
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/AgentProfileList" } } } }
    post:
      tags: [agent-profiles]
      operationId: createAgentProfile
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AgentProfileCreate" } } } }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/AgentProfile" } } } }
  /agent-profiles/{agent_profile_id}:
    parameters: [{ $ref: "#/components/parameters/AgentProfileId" }]
    get:
      tags: [agent-profiles]
      operationId: getAgentProfile
      parameters:
        - name: include
          in: query
          schema: { type: string, example: "presence,current_run" }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/AgentProfile" } } } }
    patch:
      tags: [agent-profiles]
      operationId: updateAgentProfile
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AgentProfilePatch" } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/AgentProfile" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /agent-profiles/{agent_profile_id}/commands/enable:
    parameters: [{ $ref: "#/components/parameters/AgentProfileId" }]
    post:
      tags: [agent-profiles]
      operationId: enableAgentProfile
      description: 允许被调度；不等于立刻运行。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/AgentProfile" } } } }
  /agent-profiles/{agent_profile_id}/commands/disable:
    parameters: [{ $ref: "#/components/parameters/AgentProfileId" }]
    post:
      tags: [agent-profiles]
      operationId: disableAgentProfile
      description: 停止新调度；活动 Run 按策略 interrupt（默认）。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: false, content: { application/json: { schema: { $ref: "#/components/schemas/DisableOptions" } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/AgentProfile" } } } }
  /workspaces/{workspace_id}/work-items:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [work-items]
      operationId: listWorkItems
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: priority, in: query, schema: { type: string } }
        - { name: assignee, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemList" } } } }
    post:
      tags: [work-items]
      operationId: createWorkItem
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemCreate" } } } }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
  /work-items/{work_item_id}:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    get:
      tags: [work-items]
      operationId: getWorkItem
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
    patch:
      tags: [work-items]
      operationId: updateWorkItem
      description: 普通字段修改；status 不允许任意 PATCH。
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemPatch" } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /work-items/{work_item_id}/commands/assign:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    post:
      tags: [work-items]
      operationId: assignWorkItem
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_profile_id, expected_version]
              properties:
                agent_profile_id: { type: string }
                expected_version: { type: integer }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /work-items/{work_item_id}/commands/move:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    post:
      tags: [work-items]
      operationId: moveWorkItem
      description: 看板移动、排序与状态机校验。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, expected_version]
              properties:
                status: { $ref: "#/components/schemas/WorkItemStatus" }
                expected_version: { type: integer }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /work-items/{work_item_id}/commands/block:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    post:
      tags: [work-items]
      operationId: blockWorkItem
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, message, expected_version]
              properties:
                code: { type: string }
                message: { type: string }
                source: { type: string }
                expected_version: { type: integer }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
  /work-items/{work_item_id}/commands/unblock:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    post:
      tags: [work-items]
      operationId: unblockWorkItem
      description: 解除阻塞回到 in_progress；恢复执行由用户显式创建新 Run。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_version]
              properties:
                expected_version: { type: integer }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
  /work-items/{work_item_id}/runs:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    get:
      tags: [runs]
      operationId: listWorkItemRuns
      summary: 一个任务的全部 Run（对话轮次历史；不可覆盖的执行尝试）
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/ExecutionRunList" } } } }
    post:
      tags: [runs]
      operationId: createRun
      description: 创建 ExecutionRun；立即 202，真正开始执行由 SSE 事件确认。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }, { $ref: "#/components/parameters/IfMatch" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RunCreate" } } } }
      responses:
        "202": { content: { application/json: { schema: { $ref: "#/components/schemas/RunAccepted" } } } }
  /runs/{run_id}:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    get:
      tags: [runs]
      operationId: getRun
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/ExecutionRun" } } } }
  /runs/{run_id}/events:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    get:
      tags: [runs]
      operationId: listRunEvents
      summary: 回放 Run 事件历史（对话页消息重建；只读投影，按 run_seq 排序）
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/RunEventList" } } } }
  /runs/{run_id}/commands/input:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    post:
      tags: [runs]
      operationId: sendRunInput
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instruction]
              properties:
                instruction: { type: string }
      responses:
        "202": { description: 命令已接受 }
  /runs/{run_id}/commands/interrupt:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    post:
      tags: [runs]
      operationId: interruptRun
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses: { "202": { description: 命令已接受 } }
  /runs/{run_id}/commands/cancel:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    post:
      tags: [runs]
      operationId: cancelRun
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses: { "202": { description: 命令已接受 } }
  /runs/{run_id}/commands/retry:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    post:
      tags: [runs]
      operationId: retryRun
      description: 总是创建新 Run（retry_of=当前 run），不覆盖终态；返回完整 ExecutionRun。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "202": { content: { application/json: { schema: { $ref: "#/components/schemas/ExecutionRun" } } } }
  /runs/{run_id}/commands/resume:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    post:
      tags: [runs]
      operationId: resumeRun
      description: 仅在能力声明 resume 时可用，否则 422。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "202": { description: 命令已接受 }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /runs/{run_id}/approvals:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    get:
      tags: [approvals]
      operationId: listApprovals
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalList" } } } }
  /runs/{run_id}/approvals/{approval_id}/commands/resolve:
    parameters: [{ $ref: "#/components/parameters/RunId" }, { $ref: "#/components/parameters/ApprovalId" }]
    post:
      tags: [approvals]
      operationId: resolveApproval
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                reason: { type: string }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Approval" } } } }
  /runs/{run_id}/artifacts:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    get:
      tags: [artifacts]
      operationId: listArtifacts
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/ArtifactList" } } } }
  /workspaces/{workspace_id}/activities:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [workspace]
      operationId: listActivities
      parameters: [{ $ref: "#/components/parameters/Cursor" }]
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/ActivityList" } } } }
...
  /runtime-bindings/{runtime_binding_id}/commands/probe:
    parameters: [{ $ref: "#/components/parameters/RuntimeBindingId" }]
    post:
      tags: [runtime-bindings]
      operationId: probeRuntimeBinding
      description: 探测版本、认证、协议与能力；不启动业务 Run。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "202": { content: { application/json: { schema: { $ref: "#/components/schemas/ProbeResult" } } } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    WorkspaceId: { name: workspace_id, in: path, required: true, schema: { type: string } }
...
  responses:
    Conflict:
      description: 幂等冲突或乐观锁版本冲突
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    CursorExpired:
      description: SSE cursor 已过保留窗口；客户端需重新 bootstrap
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
  schemas:
...
    Workspace:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        timezone: { type: string }
        version: { type: integer }
    WorkspacePatch:
      type: object
      properties:
        name: { type: string }
        timezone: { type: string }
        expected_version: { type: integer }
    Bootstrap:
      type: object
      properties:
        workspace: { $ref: "#/components/schemas/Workspace" }
        dashboard: { $ref: "#/components/schemas/Dashboard" }
        agents: { $ref: "#/components/schemas/AgentProfileList" }
        work_items: { $ref: "#/components/schemas/WorkItemList" }
        health: { $ref: "#/components/schemas/SystemHealth" }
        event_cursor: { type: integer }
...